docs(examples): fold event filter into dsl example
